Finalizado

+++ PracticaINDIVIDUAL.API/Commands/FuncionCMD/FuncionCommand.cs
@@ -39,18 +39,18 @@ namespace PracticaINDIVIDUAL.API.Commands.FuncionCMD
             funcion.Pelicula = pelicula;
             funcion.SalaId = request.SalaId;
             var sala = _context.Salas.Find(request.SalaId) ?? throw new ElementNotFoundException("Sala no encontrada");
-            funcion.Sala = new Sala { Nombre = sala.Nombre, Capacidad = sala.Capacidad };
+            funcion.Sala = sala;
             funcion.Tickets = new List<Ticket>(sala.Capacidad);
             funcion.Horario = request.Horario;
             var horarioProximo = request.Horario.Add(new TimeSpan(2, 30, 0));
             var horarioAnterior = request.Horario.Add(new TimeSpan(-2, 30, 0));
             var genero = _context.Generos.Find(pelicula.GeneroId) ?? throw new ElementNotFoundException("Género no encontrado");
  
-                if (funciones.Any(f => f.Fecha == funcion.Fecha && f.Horario == funcion.Horario && f.SalaId == funcion.SalaId))
+                if (funciones.Any(f => f.Fecha.ToString("yyyy-MM-dd") == funcion.Fecha.ToString("yyyy-MM-dd") && f.Horario == funcion.Horario && f.SalaId == funcion.SalaId))
                 {
                     throw new ElementAlreadyExistException("Ya existe una función para ese día y horario en esa sala");
                 }
-                if(funciones.Any(f => f.Horario <= horarioProximo && f.Horario >= horarioAnterior && f.SalaId == funcion.SalaId))
+                if(funciones.Any(f => f.Horario <= horarioProximo && f.Horario >= horarioAnterior && f.SalaId == funcion.SalaId && f.Fecha.ToString("yyyy-MM-dd") == funcion.Fecha.ToString("yyyy-MM-dd")))
                 {
                     throw new ElementAlreadyExistException("No se puede crear una función para ese día y horario en esa sala, hay superposición horaria. Intente nuevamente");
                 }
@@ -80,26 +80,24 @@ namespace PracticaINDIVIDUAL.API.Commands.FuncionCMD
                     Horario = funcion.Horario.ToString()
                 };
             }
-            
-        }
 
         public async Task<TicketDTOResponseTickets> crearTicketFuncion(int id, TicketDTO request)
         {
             List<TicketDTOResponseIDTicket> ticketsVendidos = new List<TicketDTOResponseIDTicket>();
-            var funcion = _context.Funciones.Find(id);
-            var pelicula = _context.Peliculas.Find(funcion.PeliculaId);
-            var genero = _context.Generos.Find(pelicula.GeneroId);
-            var sala = _context.Salas.Find(funcion.SalaId);
-            if (funcion == null)
+            var funcion = _context.Funciones.Find(id) ?? throw new ElementNotFoundException("No existe la función");
+            var pelicula = _context.Peliculas.Find(funcion.PeliculaId) ?? throw new ElementNotFoundException("No existe la película");
+            var genero = _context.Generos.Find(pelicula.GeneroId) ?? throw new ElementNotFoundException("No existe el género");
+            var sala = _context.Salas.Find(funcion.SalaId) ?? throw new ElementNotFoundException("No existe la sala");
+            if (request.Cantidad > sala.Capacidad && sala.Capacidad != 0)
             {
-                throw new Exception("No existe la funcion");
+                throw new InvalidOperationException("No es posible realizar la operación, se supera la cantidad máxima de tickets disponibles. Intente nuevamente");
             }
-
-            for (int i = 0; i< request.Cantidad; i++)
+            for (int i = 0; i < request.Cantidad; i++)
             {
+
                 if (sala.Capacidad == 0)
                 {
-                    throw new Exception("No hay capacidad en la sala");
+                    throw new Exception("Tickets agotados para esta función");
                 }
                 var ticket = new Ticket();
                 ticket.TicketId = new Guid();
@@ -107,7 +105,7 @@ namespace PracticaINDIVIDUAL.API.Commands.FuncionCMD
                 ticket.Funcion = funcion;
                 ticket.Usuario = request.Usuario;
                 sala.Capacidad--;
-                ticketsVendidos.Add(new TicketDTOResponseIDTicket { ticketId = ticket.TicketId});
+                ticketsVendidos.Add(new TicketDTOResponseIDTicket { ticketId = ticket.TicketId });
                 _context.Add(ticket);
             }
             await _context.SaveChangesAsync();
@@ -115,7 +113,7 @@ namespace PracticaINDIVIDUAL.API.Commands.FuncionCMD
             {
                 Tickets = ticketsVendidos,
                 Usuario = request.Usuario,
-                Funcion = new FuncionDTOResponse 
+                Funcion = new FuncionDTOResponse
                 {
                     Fecha = funcion.Fecha,
                     Horario = funcion.Horario.ToString(),
@@ -131,41 +129,47 @@ namespace PracticaINDIVIDUAL.API.Commands.FuncionCMD
                             Nombre = genero.Nombre
                         }
                     },
+                    Sala = new SalaDTOResponse
+                    {
+                        SalaId = sala.SalaId,
+                        Nombre = sala.Nombre,
+                        Capacidad = sala.Capacidad
+                    }
                 }
 
             };
-            
-            
-           
-            
-            
-            
 
-        
         }
 
-        public async Task<FuncionDTOResponse> eliminarFuncion(int funcionId)
+        public async Task<FuncionDTOResponseDetail> eliminarFuncion(int funcionId)
         {
             List<Funcion> funciones = await _context.Funciones.ToListAsync();
             if (funciones == null)
             {
-                throw new Exception("No hay funciones");
+                throw new ElementNotFoundException("No hay funciones");
             }
             var funcion = await _context.Funciones.FindAsync(funcionId);
             if (funcion == null)
             {
-                throw new Exception("No existe la funcion");
+                throw new ElementNotFoundException("No existe la funcion");
+            }
+            var tickets = await _context.Tickets.Where(t => t.FuncionId == funcionId).ToListAsync();
+            if (tickets.Count > 0)
+            {
+                throw new InvalidOperationException("No se puede eliminar la función porque hay tickets vendidos");
             }
             _context.Funciones.Remove(funcion);
             await _context.SaveChangesAsync();
-            return new FuncionDTOResponse
+            return new FuncionDTOResponseDetail
             {
                 Fecha = funcion.Fecha,
                 Horario = funcion.Horario.ToString(),
                 FuncionId = funcion.FuncionId,
-            };
-                  
+            }; 
         }
-
     }
+
+
+
+    
 }
